+++ main.go
@@ -2,14 +2,145 @@ package main
 
 import (
 	"bufio"
+	"encoding/json"
 	"fmt"
 	"github.com/spf13/viper"
+	"golang.org/x/net/context"
+	"golang.org/x/oauth2"
+	"golang.org/x/oauth2/google"
+	"google.golang.org/api/sheets/v4"
+	"io/ioutil"
+	"log"
 	"math/rand"
+	"net/http"
+	"net/url"
 	"os"
+	"os/user"
+	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
 )
+
+// getClient uses a Context and Config to retrieve a Token
+// then generate a Client. It returns the generated Client.
+func getClient(ctx context.Context, config *oauth2.Config) *http.Client {
+	cacheFile, err := tokenCacheFile()
+	if err != nil {
+		log.Fatalf("Unable to get path to cached credential file. %v", err)
+	}
+	tok, err := tokenFromFile(cacheFile)
+	if err != nil {
+		tok = getTokenFromWeb(config)
+		saveToken(cacheFile, tok)
+	}
+	return config.Client(ctx, tok)
+}
+
+// getTokenFromWeb uses Config to request a Token.
+// It returns the retrieved Token.
+func getTokenFromWeb(config *oauth2.Config) *oauth2.Token {
+	authURL := config.AuthCodeURL("state-token", oauth2.AccessTypeOffline)
+	fmt.Printf("Go to the following link in your browser then type the "+
+		"authorization code: \n%v\n", authURL)
+
+	var code string
+	if _, err := fmt.Scan(&code); err != nil {
+		log.Fatalf("Unable to read authorization code %v", err)
+	}
+
+	tok, err := config.Exchange(context.TODO(), code)
+	if err != nil {
+		log.Fatalf("Unable to retrieve token from web %v", err)
+	}
+	return tok
+}
+
+// tokenCacheFile generates credential file path/filename.
+// It returns the generated credential path/filename.
+func tokenCacheFile() (string, error) {
+	usr, err := user.Current()
+	if err != nil {
+		return "", err
+	}
+	tokenCacheDir := filepath.Join(usr.HomeDir, ".credentials")
+	os.MkdirAll(tokenCacheDir, 0700)
+	return filepath.Join(tokenCacheDir,
+		url.QueryEscape("sheets.googleapis.com-go-quickstart.json")), err
+}
+
+// tokenFromFile retrieves a Token from a given file path.
+// It returns the retrieved Token and any read error encountered.
+func tokenFromFile(file string) (*oauth2.Token, error) {
+	f, err := os.Open(file)
+	if err != nil {
+		return nil, err
+	}
+	t := &oauth2.Token{}
+	err = json.NewDecoder(f).Decode(t)
+	defer f.Close()
+	return t, err
+}
+
+// saveToken uses a file path to create a file and store the
+// token in it.
+func saveToken(file string, token *oauth2.Token) {
+	fmt.Printf("Saving credential file to: %s\n", file)
+	f, err := os.OpenFile(file, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0600)
+	if err != nil {
+		log.Fatalf("Unable to cache oauth token: %v", err)
+	}
+	defer f.Close()
+	json.NewEncoder(f).Encode(token)
+}
+
+func readGoogleSheetsBooklist() []string {
+	ctx := context.Background()
+
+	b, err := ioutil.ReadFile("conf/client_secret.json")
+	if err != nil {
+		log.Fatalf("Unable to read client secret file: %v", err)
+	}
+
+	// If modifying these scopes, delete your previously saved credentials
+	// at ~/.credentials/sheets.googleapis.com-go-quickstart.json
+	config, err := google.ConfigFromJSON(b, "https://www.googleapis.com/auth/spreadsheets.readonly")
+	if err != nil {
+		log.Fatalf("Unable to parse client secret file to config: %v", err)
+	}
+	client := getClient(ctx, config)
+
+	srv, err := sheets.New(client)
+	if err != nil {
+		log.Fatalf("Unable to retrieve Sheets Client %v", err)
+	}
+
+	// Prints the names and majors of students in a sample spreadsheet:
+	// https://docs.google.com/spreadsheets/d/1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms/edit
+	spreadsheetId := "1mFAiopHeUfjyeMjVfDFxKGqSh19jBsTrSN5hEO6ppGs"
+	readRange := "Suggestions!A2:E"
+	resp, err := srv.Spreadsheets.Values.Get(spreadsheetId, readRange).Do()
+	if err != nil {
+		log.Fatalf("Unable to retrieve data from sheet. %v", err)
+	}
+
+	if len(resp.Values) > 0 {
+		return formatGoogleSheetsBookList(resp.Values)
+	} else {
+		fmt.Print("No data found.")
+	}
+	return []string{}
+}
+
+func formatGoogleSheetsBookList(i [][]interface{}) []string {
+	gsBookList := []string{}
+	for _, row := range i {
+		// Print columns A and E, which correspond to indices 0 and 4.
+		bookStr := fmt.Sprintf("%s", row[1])
+		gsBookList = append(gsBookList, fmt.Sprintf("%s: %s", row[0], strings.Replace(bookStr, ":", "-", -1)))
+	}
+	return gsBookList
+}
 
 var (
 	reader    = bufio.NewReader(os.Stdin)
@@ -21,7 +152,6 @@ var (
 )
 
 func main() {
-
 	// Clear the screen.
 	for i := 0; i < 100; i++ {
 		fmt.Println()
@@ -37,7 +167,7 @@ func main() {
 		if display {
 			printBookVotes()
 			fmt.Print("\nSelect one of the following options: \n\na/A: Add the selections of a voter. " +
-				"\nr/R: Read in the list of books. \nselect: Select a book randomly based on the " +
+				"\nr/R: Read in the list of books. \nw/W: Update book list from Google Sheets. \nselect: Select a book randomly based on the " +
 				"given votes and finish program (select must be typed completely) \nx: Exit Book " +
 				"Selector.\n\nWhat would you like do next? \n")
 		}
@@ -53,6 +183,9 @@ func main() {
 			display = true
 		case "R":
 			readInBooks()
+			display = true
+		case "W":
+			writeBooks()
 			display = true
 		case "SELECT":
 			display = false
@@ -208,6 +341,24 @@ func readInBooks() {
 
 	fmt.Println("\nBOOKS HAVE BEEN READ IN")
 
+}
+
+func writeBooks() {
+	file, err := os.Create(viper.GetString("booklistlocation"))
+	if err != nil {
+		fmt.Printf("Error writing books from this file: %s | %s", viper.GetString("booklistlocation"), err)
+	}
+	defer file.Close()
+
+	w := bufio.NewWriter(file)
+	lines := readGoogleSheetsBooklist()
+	for _, line := range lines {
+		fmt.Fprintln(w, line)
+	}
+
+	if err = w.Flush(); err != nil {
+		panic(err)
+	}
 }
 
 func printBookVotes() {
